Implements the extends keyword for types defined in the dsl.
Fixes issue with test for extends.

Working to get the tests to pass.

The nools parser test for defines uses a deep equal to a known
structure.  There is a new property in the define parse output, extend:
<typename> , this is null unless there is an actual base class.  Adding
null, trivial change for this.

Implements the extends keyword for types defined in the dsl.